fix: removed init
- Renamed cmd package to cli package.
- Removed init function usage in cli package.
- Updated README.md.
