+++ main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"github.com/strongishllama/cmfive-cli/pkg/cmd"
+	"github.com/strongishllama/cmfive-cli/pkg/cli"
 	"github.com/strongishllama/cmfive-cli/pkg/cmfive"
 )
 
 func main() {
 	cmfive.TemplatesDir = "cmfive/"
-	cmd.Execute()
+	cli.Execute()
 }
