--- pkg/cmd/root.go
@@ -1,93 +1,0 @@
-package cmd
-
-import (
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var cfgFile string
-
-var rootCmd = &cobra.Command{
-	Use:     "cmfive <command> <subcommand> [flags]",
-	Short:   "A CLI tool to help create Cmfive applications.",
-	Example: "cmfive new module payroll",
-}
-
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		os.Exit(1)
-	}
-}
-
-func init() {
-	// cobra.OnInitialize(initConfig)
-
-	rootCmd.SetHelpFunc(func(cmd *cobra.Command, args []string) {
-		if rootCmd.Long != "" {
-			cmd.Printf("%s\n\n", rootCmd.Long)
-		} else {
-			cmd.Printf("%s\n\n", rootCmd.Short)
-		}
-
-		cmd.Println("USAGE:")
-		cmd.Printf("  %s\n\n", rootCmd.Use)
-
-		commands := rootCmd.Commands()
-		if len(commands) != 0 {
-			cmd.Println("COMMANDS:")
-
-			for _, c := range commands {
-				cmd.Printf("  %s\t\t%s\n", c.Name(), c.Short)
-			}
-		}
-		cmd.Println()
-
-		flagUsages := rootCmd.LocalFlags().FlagUsages()
-		if flagUsages != "" {
-			cmd.Println("FLAGS:")
-			cmd.Println(flagUsages)
-		}
-
-		cmd.Println("EXAMPLE:")
-		cmd.Printf("  %s\n", rootCmd.Example)
-	})
-
-	// Here you will define your flags and configuration settings.
-	// Cobra supports persistent flags, which, if defined here,
-	// will be global for your application.
-
-	// rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.cmfive-cli.yaml)")
-
-	// Cobra also supports local flags, which will only run
-	// when this action is called directly.
-	// rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
-
-// initConfig reads in config file and ENV variables if set.
-// func initConfig() {
-// 	if cfgFile != "" {
-// 		// Use config file from the flag.
-// 		viper.SetConfigFile(cfgFile)
-// 	} else {
-// 		// Find home directory.
-// 		home, err := homedir.Dir()
-// 		if err != nil {
-// 			fmt.Println(err)
-// 			os.Exit(1)
-// 		}
-
-// 		// Search config in home directory with name ".cmfive-cli" (without extension).
-// 		viper.AddConfigPath(home)
-// 		viper.SetConfigName(".cmfive-cli")
-// 	}
-
-// 	viper.AutomaticEnv() // read in environment variables that match
-
-// 	// If a config file is found, read it in.
-// 	if err := viper.ReadInConfig(); err == nil {
-// 		fmt.Println("Using config file:", viper.ConfigFileUsed())
-// 	}
-// }
